% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheck.R
\name{datacheck}
\alias{datacheck}
\title{Check Input Data Tyoe}
\usage{
datacheck(data, label, time)
}
\arguments{
\item{data}{The input matrix of observed values. Rows represent repititions or samples, and columns represent observed time points}

\item{label}{The input vector of labels. ith element of this vector should correspond to the label of ith row of data matrix X.}

\item{time}{An input time grid vector}
}
\value{
NULL is appropriate, otherwise warning
}
\description{
\code{datacheck()} is to check whether the data to be input is appropriate
}
\examples{
 library(refund)
 data(DTI)
 X <- DTI$cca
 y <- DTI$case
 t <- seq(0, 1, length.out = ncol(X))
  # not run
  # datacheck(X, y, t)
 allData <- cbind(X, y)
 allData <- allData[which(DTI$visit == 1), ]
 allData <- na.omit(allData)
 y <- allData[, ncol(allData)]
 X <- allData[, -ncol(allData)]
 t <- seq(0, 1, length.out = ncol(X))
 datacheck(X, y, t)
}
