% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findweights.R
\name{findweights}
\alias{findweights}
\title{Find optimal weights based on the quantileloss}
\usage{
findweights(weight, scorematrixlist, par.sig = 1000)
}
\arguments{
\item{weight}{The initial weight to be optimized}

\item{scorematrixlist}{A list of quantileloss matrices}

\item{par.sig}{The parameter in sigmoid function, default is 1000}
}
\value{
It returns a list, which is the output of \code{optim()} function. Details can be seen in \code{\link{optim}}.
}
\description{
A weight is multiplied to the quantileloss to select the important projection scores.
}
