% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOBJ.R
\name{FOBJ}
\alias{FOBJ}
\title{Functional Objection}
\usage{
FOBJ(X, y, t, optns = list(smooth = FALSE, FVEthreshold = 99.99, a0 =
  0.02, M0 = 10, par.sig = 1000), weighted = FALSE)
}
\arguments{
\item{X}{The input matrix of observed values. Rows represent repititions or samples, and columns represent observed time points.}

\item{y}{The input vector of labels. ith element of this vector should correspond to the label of ith row of data matrix X.}

\item{t}{An input time grid vector.}

\item{optns}{A list to control computing such as smoothing, threshold and quantile range.}

\item{weighted}{Whether weight for quantile loss is to be used. \code{FALSE} is default.}
}
\value{
It returns a class "FOBJ".
\item{DataGroupby}{The list group by the label of each group}
\item{y}{Vector. Label of each group}
\item{t}{Original time points of observation}
\item{K}{Numeber. How many groups}
\item{phi}{Matrix. The functional principle componants, listed by colums}
\item{scores}{List. The projection scores of each group}
\item{cutPoints}{Numebr. How many componants}
\item{quantlist}{List. The quantile of projection scores of each group}
\item{tau}{Vector. Quantile counter}
\item{weight}{Matrix. Weight for classification. Default is matrix 1. The number of colums is equal to \code{tau}}
\item{optns}{List. The same to the original optns}
\item{weighted}{Logical. The same to the parameter "weighted"}
\item{propotion}{Vector. Contains the proportion of variance of each FPC}
}
\description{
\code{FOBJ()} is used to compute imformation for within FDA framework.
}
\examples{
rm(list = ls())
library(refund)
data(DTI)
X <- DTI$cca
y <- DTI$case
t <- seq(0, 1, length.out = ncol(X))
#not run
#datacheck(X, y, t)
allData <- cbind(X, y)
allData <- allData[which(DTI$visit == 1), ]
allData <- na.omit(allData)
y <- allData[, ncol(allData)]
X <- allData[, -ncol(allData)]
t <- seq(0, 1, length.out = ncol(X))
w1 <- FOBJ(X, y, t)
}
