% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhi.R
\name{getPhi}
\alias{getPhi}
\title{Get functional priciple componants}
\usage{
getPhi(Cov, FVEthreshold = 99.99, t)
}
\arguments{
\item{Cov}{Covariance surface}

\item{FVEthreshold}{FVE}

\item{t}{Time points}
}
\description{
FPCA
}
\references{
Yao F, Müller H G, Wang J L. Functional data analysis for sparse longitudinal data[J]. Journal of the American Statistical Association, 2005, 100(470): 577-590.
}
