% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXi.R
\name{getXi}
\alias{getXi}
\title{Get projection scores from the estimated eigenfunctions}
\usage{
getXi(X.curve, phi, t)
}
\arguments{
\item{X.curve}{The input matrix of observed values. Rows represent repititions or samples, and columns represent observed time points}

\item{phi}{A matrix containing normalized eigenfunctions, which is always generated by \code{FOBJ}}

\item{t}{A input time grid vector}
}
\value{
It returns a matrix:
\item{Xi}{A matrix containing the projection scores}
}
\description{
After getting estimated eigenfunctions, this function can be used to compute projection scores of observed functional data.
}
