% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.FOBJ.R
\name{plot.FOBJ}
\alias{plot.FOBJ}
\title{Visulization of FOBJ}
\usage{
\method{plot}{FOBJ}(x, J = 4)
}
\arguments{
\item{x}{An FOBJ, generated by \code{FOBJ}}

\item{J}{The number of PCs to be plotted. Only positive number is allowed}
}
\description{
This function is aimed to plot the functional data from different aspects
}
