% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooledCov.R
\name{pooledcov}
\alias{pooledcov}
\title{Pooled Covariance Function}
\usage{
pooledcov(Xlist, t)
}
\arguments{
\item{Xlist}{A list of gruoped functional data.}

\item{t}{Time points}
}
\description{
Pooled Covariance Function
}
