% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.quantile.R
\name{pre.quantile}
\alias{pre.quantile}
\title{Predict Function Using Quantile}
\usage{
pre.quantile(X.curve, obj, weighted = obj$weighted)
}
\arguments{
\item{X.curve}{A vector denotes the discrete points of a single curve.}

\item{obj}{An FOBJ object containing the information for classification.}

\item{weighted}{Whether multiple weight by \code{FOBJ} is to be used or not. Default the same to \code{obj$weighted}.}
}
\description{
Assign label to a single curve
}
