% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileloss.R
\name{quantileloss}
\alias{quantileloss}
\title{Computing Quantileloss}
\usage{
quantileloss(Xi, q, tau)
}
\arguments{
\item{Xi}{A vector/matrix of projection score value for ith curve}

\item{q}{Emperical quantile}

\item{tau}{Single tau}
}
\description{
Compute quantile loss for a single curve
}
\references{
To be confirmed
}
