% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid Function}
\usage{
sigmoid(x, par.sig = 1000)
}
\arguments{
\item{x}{vectors or values to be assigned}

\item{par.sig}{parameter for sigmoid function}
}
\value{
0 or 1, based on whether your input number or vector is positive or negative.
}
\description{
Using Sigmoid function to smoothly assign a label
}
\examples{
a <- runif(10, -1, 1)
b <- sigmoid(a)
}
